Revised savemap into Explorer svc
